# wafaa/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MusicBox API",
        "description": "API pour gérer les artistes, albums et chansons",
        "contact": {
            "email": "admin@musicbox.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur Local"
        }
    ],
    "paths": {
...
        "/api/songs": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Liste des chansons",
                "operationId": "0388bb28c35eec49b8b492d169ce3575",
                "parameters": [
                    {
                        "name": "min_duration",
                        "in": "query",
                        "description": "Durée minimale (en secondes)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "max_duration",
                        "in": "query",
                        "description": "Durée maximale (en secondes)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Recherche par titre",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des chansons"
                    }
                }
            },
            "post": {
                "tags": [
                    "Songs"
                ],
                "summary": "Créer une chanson",
                "operationId": "4d2ab730930b4397ceb03cd2a237e01c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "duration",
                                    "album_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Shape of You"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 210
                                    },
                                    "album_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chanson créée avec succès"
                    },
                    "400": {
                        "description": "Requête invalide"
                    }
                }
            }
        },
        "/api/songs/{id}": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Afficher une chanson",
                "operationId": "92fbec836108444814c7d79d785f3895",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chanson trouvée"
                    },
                    "404": {
                        "description": "Chanson non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Songs"
                ],
                "summary": "Mettre à jour une chanson",
                "operationId": "72f2f578765c2b6dcf0950443c8d24e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Perfect"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 240
                                    },
                                    "album_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chanson mise à jour"
                    },
                    "404": {
                        "description": "Chanson non trouvée"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Songs"
                ],
                "summary": "Supprimer une chanson",
                "operationId": "81db80566e30283003f122862b3bd3ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chanson supprimée"
                    },
                    "404": {
                        "description": "Chanson non trouvée"
                    }
                }
            }
        },
        "/api/songs/search": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Rechercher des chansons par titre ou artiste",
                "operationId": "cc4c089110a4aa64f9711bcc9f30ca34",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Mot-clé (titre ou artiste)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultats trouvés"
                    },
                    "404": {
                        "description": "Aucun résultat trouvé"
                    }
                }
            }
        }
    },
...
}